Update database host IP in alembic config
The alembic.ini configuration file has been updated to reflect the change in the database server's IP address. The previous IP (172.19.0.4) has been replaced with the new IP (192.168.1.28) to ensure that Alembic can connect to the correct database host for migrations. This change is necessary to maintain the connection to the database after the server's network configuration has been updated.